requirements.txt
alembic==1.0.11
aniso8601==7.0.0
appdirs==1.4.3
attrs==19.1.0
autopep8==1.4.4
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
requests-cache==0.5.1
six==1.12.0
SQLAlchemy==1.3.6
toml==0.10.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
yapf==0.28.0
